I znowu masa zmian

+++ app/src/main/java/com/dyszlewskiR/edu/scientling/data/file/WordFileSystem.java
@@ -16,42 +16,42 @@ import java.io.IOException;
  * Klasa stanowi dodatkową warstwę zapisu pliku, na której można wykonywać dodatkowe działania,
  * takie jak przeskalowanie obrazków
  */
-
 public class WordFileSystem {
 
     private static final int MAX_IMAGE_SIZE = 300;
 
     public static void saveImage(String fileName, String setCatalog, byte[] data, boolean resize, Context context) throws IOException {
-        if(resize ){
+        if (resize) {
             data = getResizedData(data, MAX_IMAGE_SIZE);
         }
         //TODO zrobić wykrywanie z jakiej pamięci korzystamy w tej chwili i zapisanie w odpowiednim miejsu
-        FileUtils.saveFileInternalStorage(setCatalog, fileName,data,context);
+        FileUtils.saveFileInternalStorage(setCatalog, fileName, data, context);
     }
 
     /**
      * Metoda zwracająca dane zmniejszonego obrazka
-     * @param data
-     * @param maxSize
-     * @return
+     *
+     * @param data    dane orginalnego obrazka
+     * @param maxSize najdłuższy możliwy wymiar obrazka wyrażony w pikselach.
+     * @return dane pomniejszonego obrazka
      */
-    private static byte[] getResizedData(byte[] data, int maxSize){
+    private static byte[] getResizedData(byte[] data, int maxSize) {
         Bitmap bitmap = BitmapUtils.getBitmap(data);
         //w przypadku obrazka mniejszego od rozmiaru do którego chcieliśmy konwertować zwracamy orginalne dane
-        if(bitmap.getWidth() < maxSize && bitmap.getHeight() < maxSize){
+        if (bitmap.getWidth() < maxSize && bitmap.getHeight() < maxSize) {
             return data;
         }
-        bitmap = BitmapUtils.resize(bitmap, maxSize,false);
+        bitmap = BitmapUtils.resize(bitmap, maxSize, false);
         return BitmapUtils.toByteArray(bitmap);
     }
 
-    public static void saveImage(String fileName, String setCatalog,  Uri uri, Context context, boolean resize) throws IOException {
+    public static void saveImage(String fileName, String setCatalog, Uri uri, Context context, boolean resize) throws IOException {
         byte[] data = UriUtils.toByteArray(uri, context);
         saveImage(fileName, setCatalog, data, resize, context);
     }
 
 
-    public static void saveRecord(String fileName, String setCatalog, byte[] data, Context context) throws IOException{
+    public static void saveRecord(String fileName, String setCatalog, byte[] data, Context context) throws IOException {
         //TODO sprawdzenie w jakiej pamięci powinno zostać zapisane nagranie
         FileUtils.saveFileInternalStorage(setCatalog, fileName, data, context);
     }
@@ -61,16 +61,16 @@ public class WordFileSystem {
         saveRecord(fileName, setCatalog, data, context);
     }
 
-    public static Uri getImageUri(String fileName, String setCatalog, Context context ){
+    public static Uri getImageUri(String fileName, String setCatalog, Context context) {
         //TODO pobierać z odpowiedniej pamięcia
-        if(fileName == null || setCatalog == null){
+        if (fileName == null || setCatalog == null) {
             return null;
         }
         return FileUtils.getInternalStorageUri(fileName, setCatalog, context);
     }
 
-    public static Uri getRecordUri(String fileName, String setCatalog, Context context){
-        if(fileName == null || setCatalog == null){
+    public static Uri getRecordUri(String fileName, String setCatalog, Context context) {
+        if (fileName == null || setCatalog == null) {
             return null;
         }
         return FileUtils.getInternalStorageUri(fileName, setCatalog, context);
@@ -78,16 +78,28 @@ public class WordFileSystem {
 
 
     // TODO można zrobić jedną metodę dla obrazków i nagrań
-    public static void deleteImage(String fileName, String secCatalog, Context context){
-        FileUtils.deleteFileInternalStorage(fileName, secCatalog, context);
+    public static void deleteImage(String fileName, String setCatalog, Context context) {
+        if (fileName != null && setCatalog != null) {
+            FileUtils.deleteFileInternalStorage(fileName, setCatalog, context);
+        }
     }
 
-    public static void deleteRecord(String fileName, String setCatalog, Context context){
-        FileUtils.deleteFileInternalStorage(fileName, setCatalog, context);
+    public static void deleteRecord(String fileName, String setCatalog, Context context) {
+        if (fileName != null && setCatalog != null) {
+            FileUtils.deleteFileInternalStorage(fileName, setCatalog, context);
+        }
     }
 
-    public static void deleteCatalog(String catalogName, Context context){
-        FileUtils.deleteDirectory(catalogName,  context);
+    public static void deleteCatalog(String catalogName, Context context) {
+        FileUtils.deleteDirectory(catalogName, context);
+    }
+
+    public static boolean checkFileExist(String fileName, String catalog, Context context) {
+        if (fileName == null || catalog == null) {
+            return false;
+        }
+        boolean found = FileUtils.checkFileExist(fileName, context.getFilesDir().getAbsolutePath() + "/" + catalog);
+        return found;
     }
 
 }
